move the group transformer to the separate module
